+++ Swiftfin/Views/ItemView/ScrollViews/iPadOSCinematicScrollView.swift
@@ -164,10 +164,8 @@ extension ItemView.iPadOSCinematicScrollView {
                     Spacer()
 
                     VStack(spacing: 10) {
-                        if viewModel.item.type == .person || viewModel.item.type == .musicArtist,
-                           let typeViewModel = viewModel as? CollectionItemViewModel
-                        {
-                            ImageView(typeViewModel.item.imageSource(.primary, maxWidth: 200))
+                        if viewModel.item.type == .person || viewModel.item.type == .musicArtist {
+                            ImageView(viewModel.item.imageSource(.primary, maxWidth: 200))
                                 .failure {
                                     SystemImageContentView(systemName: viewModel.item.systemImage)
                                 }
